+++ src/cli/lib/error-out.mjs
@@ -1,0 +1,6 @@
+const errorOut = (msg, code) => {
+  process.stderr.write(msg)
+  process.exit(code) // eslint-disable-line no-process-exit
+}
+
+export { errorOut }
